+++ docker/scripts/qgis-docker-entrypoint.sh
@@ -23,7 +23,6 @@ copy_qgis_configuration() {
     export QGIS_OPTIONS_PATH=$HOME
 
     # Create plugin dir
-    export QGIS_PLUGINPATH=$QJAZZ_VOLUME/plugins
     mkdir -p $QGIS_PLUGINPATH
 }